refactor CLI for MVC
